Move non-page-tracking to end of body
