+++ website/src/components/Analytics/types.ts
@@ -1,0 +1,10 @@
+export interface HubSpotSubmittedFormData {
+  type: string;
+  eventName: string;
+  redirectUrl: string;
+  conversionId: string;
+  formGuid: string;
+  submissionValues: {
+    [key: string]: string;
+  };
+}
